[SHIPA-2020] Bugfix: traefik doesn't work with more than one https cname
