+++ internal/templates/traefik/yamls/http-ingress-route.yaml
@@ -1,0 +1,37 @@
+{{- if .Values.app.isAccessible }}
+{{- if .Values.app.ingress.http }}
+apiVersion: traefik.containo.us/v1alpha1
+kind: IngressRoute
+metadata:
+  name: {{ $.Values.app.name }}-http-ingressroute
+  annotations:
+    {{- if .Values.ingressController.className }}
+    kubernetes.io/ingress.class: {{ .Values.ingressController.className | quote }}
+    {{- end }}
+    {{- if .Values.ingressController.clusterIssuer }}
+    cert-manager.io/cluster-issuer: {{ .Values.ingressController.clusterIssuer | quote }}
+    {{- end }}
+  labels:
+    {{ $.Values.app.group }}/app-name: {{ $.Values.app.name | quote }}
+spec:
+  entryPoints:
+    - web
+  routes:
+  {{- range $_, $cname := .Values.app.ingress.http }}
+  - match: Host("{{ $cname }}")
+    kind: Rule
+    services:
+    {{- range $_, $deployment := $.Values.app.deployments }}
+    {{- range $_, $process := $deployment.processes }}
+    {{- if $process.routable }}{{- if gt $deployment.routingSettings.weight 0.0}}
+    - name: {{ printf "%s-%s-%v" $.Values.app.name $process.name $deployment.version }}
+      port: {{ $process.publicServicePort }}
+      weight: {{$deployment.routingSettings.weight}}
+      {{- end }}
+      {{- end }}
+      {{- end }}
+  {{- end }}
+  {{- end }}
+---
+{{- end }}
+{{- end }}
